requirements.txt
opencv-python
numpy
opencv-contrib-python
pillow
pylint
autopep8